+++ myems-admin/app/config.router.js
@@ -1101,6 +1101,7 @@ app
                                                     'app/services/settings/microgrid/microgridpowerconversionsystem.service.js',
                                                     'app/services/settings/microgrid/microgridsensor.service.js',
                                                     'app/services/settings/microgrid/microgridwindturbine.service.js',
+                                                    'app/services/settings/microgrid/microgriduser.service.js',
                                                     'app/services/settings/costcenter/costcenter.service.js',
                                                     'app/services/settings/contact/contact.service.js',
                                                     'app/services/settings/datasource/point.service.js',
@@ -1117,6 +1118,7 @@ app
                                                     'app/controllers/settings/microgrid/microgridpowerconversionsystem.controller.js',
                                                     'app/controllers/settings/microgrid/microgridsensor.controller.js',
                                                     'app/controllers/settings/microgrid/microgridwindturbine.controller.js',
+                                                    'app/controllers/settings/microgrid/microgriduser.controller.js',
                                                 ]
                                             }]);
                                         }